Avoid boxing: Replace Integer by int in fields
An int uses 4 bytes, an Integer uses 16 to 24 bytes on
Hotspot (http://stackoverflow.com/a/9037808/53974). That alone is not a
necessarily a reason, but boxed integers allow for a couple of
bugs (they can be null, and you never want to use `==` on them, see
http://stackoverflow.com/a/5199425/53974 and
http://stackoverflow.com/a/5199559/53974).

+++ app/src/test/java/com/kamron/pogoiv/logic/PokeSpamTest.java
@@ -13,45 +13,45 @@ public class PokeSpamTest {
     @Test
     public void testGetHowMuchWeCanEvolve() throws Exception {
         PokeSpam pokeSpamCalculator = new PokeSpam(536, 11);
-        assertEquals((int) pokeSpamCalculator.getTotalEvolvable(), 53);
+        assertEquals(pokeSpamCalculator.getTotalEvolvable(), 53);
     }
 
     @Test
     public void testEvolveRows() throws Exception {
         PokeSpam pokeSpamCalculator = new PokeSpam(536, 12);
-        assertEquals((int) pokeSpamCalculator.getEvolveRows(), 16);
+        assertEquals(pokeSpamCalculator.getEvolveRows(), 16);
     }
 
     @Test
     public void testGetEvolveExtra() throws Exception {
         PokeSpam pokeSpamCalculator = new PokeSpam(536, 12);
-        assertEquals((int) pokeSpamCalculator.getEvolveExtra(), 0);
+        assertEquals(pokeSpamCalculator.getEvolveExtra(), 0);
     }
 
     @Test
     public void testNotEnoughCandy() throws Exception {
         PokeSpam pokeSpamCalculator = new PokeSpam(2, 12);
-        assertEquals((int) pokeSpamCalculator.getTotalEvolvable(), 0);
-        assertEquals((int) pokeSpamCalculator.getEvolveRows(), 0);
-        assertEquals((int) pokeSpamCalculator.getEvolveExtra(), 0);
+        assertEquals(pokeSpamCalculator.getTotalEvolvable(), 0);
+        assertEquals(pokeSpamCalculator.getEvolveRows(), 0);
+        assertEquals(pokeSpamCalculator.getEvolveExtra(), 0);
     }
 
     @Test
     public void testBlaisorbladeTestCases() throws Exception {
-        assertEquals((new PokeSpam(11, 12).getTotalEvolvable()), (Integer) 0);
-        assertEquals((new PokeSpam(12, 12).getTotalEvolvable()), (Integer) 1);
-        assertEquals((new PokeSpam(22, 12).getTotalEvolvable()), (Integer) 1);
-        assertEquals((new PokeSpam(23, 12).getTotalEvolvable()), (Integer) 2);
+        assertEquals(new PokeSpam(11, 12).getTotalEvolvable(), 0);
+        assertEquals(new PokeSpam(12, 12).getTotalEvolvable(), 1);
+        assertEquals(new PokeSpam(22, 12).getTotalEvolvable(), 1);
+        assertEquals(new PokeSpam(23, 12).getTotalEvolvable(), 2);
     }
 
     @Test
     public void testVenonat() throws Exception {
-        assertEquals((new PokeSpam(150, 50).getTotalEvolvable()), (Integer) 3);
+        assertEquals(new PokeSpam(150, 50).getTotalEvolvable(), 3);
     }
 
     @Test
     public void testExtraPokemonBasedOnExtraCandy() throws Exception {
-        assertEquals((new PokeSpam(144, 12).getTotalEvolvable()), (Integer) 13);
+        assertEquals(new PokeSpam(144, 12).getTotalEvolvable(), 13);
     }
 
 }
